+++ Assets/Scripts/ScriptableObjects/Event/Base/GenericEventChannelSO.cs
@@ -2,6 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using System;
+using UnityEditor;
 
 namespace EventSO
 {
@@ -11,18 +12,25 @@ namespace EventSO
         public void AddEventListener(Action<T> Handler);
         public void RemoveEventListener(Action<T> Handler);
     }
+
     public abstract class GenericEventChannelSO<T> : ScriptableObject, IEventDataSO<T>
     {
         public Action<T> OnEvent;
 
 #if UNITY_EDITOR
-        public T _testData;
+        [HideInInspector] public T _testData;
+
+        public void Test()
+        {
+            OnEvent?.Invoke(_testData);
+        }
 #endif
 
         public void Invoke(T data)
         {
             OnEvent?.Invoke(data);
         }
+
         public void AddEventListener(Action<T> Handler)
         {
             OnEvent += Handler;
