+++ Assets/Scripts/ScriptableObjects/Event/Editor/EventItemEditor.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEditor;
+using UnityEngine;
+
+namespace EventSO
+{
+    [CustomEditor(typeof(EventItemSO), editorForChildClasses: true)]
+    public class EventItemEditor : GenericEventEditor<ItemSO>
+    {
+    }
+
+}
