+++ Assets/Scripts/ScriptableObjects/Event/Editor/EventItemEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c281e26ca49440a4e86886c1ef8c8d39
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
