+++ Assets/Scripts/ScriptableObjects/Event/Editor/GenericEventEditor.cs
@@ -1,0 +1,41 @@
+using EventSO;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEditor;
+using UnityEngine;
+
+
+namespace EventSO
+{
+    public abstract class GenericEventEditor<T> : Editor
+    {
+        SerializedProperty _testData;
+
+        private void OnEnable()
+        {
+            _testData = serializedObject.FindProperty("_testData");
+        }
+
+        public override void OnInspectorGUI()
+        {
+            base.OnInspectorGUI();
+
+            serializedObject.Update();
+            GUI.enabled = Application.isPlaying;
+
+            GenericEventChannelSO<T> e = target as GenericEventChannelSO<T>;
+
+            if (Application.isPlaying)
+                EditorGUILayout.PropertyField(_testData);
+
+            if (GUILayout.Button("Raise"))
+                e.Invoke(e._testData);
+
+            serializedObject.ApplyModifiedProperties();
+
+        }
+    }
+}
+
+
+
