+++ Assets/Scripts/ScriptableObjects/Event/Editor/GenericEventEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4f4da7acb93a75e4e81350ecaa86620e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
